AvrDataTypes: Fixed string new lines problem and separated the Register into a new lib
