+++ AvrDataTypes/.vscode/settings.json
@@ -3,6 +3,7 @@
         "stdint.h": "c",
         "array": "c",
         "string": "c",
-        "string_view": "c"
+        "string_view": "c",
+        "register.h": "c"
     }
 }
